+++ src/core/Stocks/View/StockOwnershipView.cs
@@ -9,7 +9,7 @@ namespace core.Stocks.View
         public Guid Id { get; set; }
         public decimal AverageCost { get; set; }
         public decimal Cost { get; set; }
-        public int Owned { get; set; }
+        public decimal Owned { get; set; }
         public string Ticker { get; set; }
         public string Category { get; set; }
         public List<Transaction> Transactions { get; set; }
